Modification pour l'enregistrement d'un model pour qu'il soit uniquement de .odt Rajout de message dans le dictionnaire Rajout en BDD le nom du fichier pour ensuite afficher le nom et non une suite de chiffre non explicite.

+++ trunk/app/Model/Modele.php
@@ -56,39 +56,50 @@ class Modele extends AppModel {
         if (isset($data['Modele']['modele']) && !empty($data['Modele']['modele'])) {
             $file = $data['Modele']['modele'];
             $success = true;
-            if (!empty($file['name'])) {
-                $this->begin();
-                $folder = WWW_ROOT . 'files/modeles';
-                $extension = strtolower(pathinfo($file['name'], PATHINFO_EXTENSION));
-                if (!empty($file['tmp_name'])) {
-                    $url = time();
-                    $success = $success && move_uploaded_file($file['tmp_name'], $folder . '/' . $url . '.' . $extension);
-                    if ($success) {
-                        $adel = $this->find('all', array('conditions' => array('formulaires_id' => $id)));
-                        foreach ($adel as $value) {
-                            //unlink($folder . '/' . $value['Modele']['fichier']);
+
+            $extension = strtolower(pathinfo($file['name'], PATHINFO_EXTENSION));
+
+            if ($extension == 'odt') {
+                if (!empty($file['name'])) {
+                    $this->begin();
+                    $folder = WWW_ROOT . 'files/modeles';
+
+                    if (!empty($file['tmp_name'])) {
+                        $url = time();
+                        $success = $success && move_uploaded_file($file['tmp_name'], $folder . '/' . $url . '.' . $extension);
+                        if ($success) {
+                            $adel = $this->find('all', array('conditions' => array('formulaires_id' => $id)));
+                            foreach ($adel as $value) {
+                                //unlink($folder . '/' . $value['Modele']['fichier']);
+                            }
+                            $this->deleteAll(array('formulaires_id' => $id));
+                            $this->create(array(
+                                'fichier' => $url . '.' . $extension,
+                                'formulaires_id' => $id,
+                                'name_fichier' => $file['name']
+                            ));
+                            $success = $success && $this->save();
                         }
-                        $this->deleteAll(array('formulaires_id' => $id));
-                        $this->create(array(
-                            'fichier' => $url . '.' . $extension,
-                            'formulaires_id' => $id
-                        ));
-                        $success = $success && $this->save();
+                    } else {
+                        $success = false;
                     }
                 } else {
                     $success = false;
                 }
-            }
 
-            if ($success) {
-                $this->commit();
+                if ($success) {
+                    $this->commit();
+                    return (0);
+                } else {
+                    $this->rollback();
+                    return (1);
+                }
             } else {
-                $this->rollback();
-                return false;
+                return (2);
             }
         }
 
-        return true;
+        return (3);
     }
 
 }
